Add lit test for cmake cmake_minimum_required
