+++ clang/test/dpct/cmake_migation/case_009/case_009.cpp
@@ -1,0 +1,11 @@
+// UNSUPPORTED: system-windows
+// RUN: rm -rf %T && mkdir -p %T
+// RUN: cd %T
+// RUN: cp %S/input.cmake ./input.cmake
+// RUN: dpct -in-root ./ -out-root out  ./input.cmake  --rule-file %S/rules.yaml --migrate-cmake-script-only --cuda-include-path="%cuda-path/include"
+// RUN: echo "begin" > %T/diff.txt
+// RUN: diff --strip-trailing-cr %S/expected.txt %T/out/input.cmake >> %T/diff.txt
+// RUN: echo "end" >> %T/diff.txt
+
+// CHECK: begin
+// CHECK-NEXT: end
