+++ clang/test/dpct/cmake_migation/case_009/expected.txt
@@ -1,0 +1,10 @@
+cmake_minimum_required(VERSION 3.24)
+cmake_minimum_required(VERSION 3.26)
+cmake_minimum_required(VERSION 3.24...3.26)
+set  (  VER 
+  3.15)
+set(MIN_VER 3.10)
+set(MAX_VER 3.26)
+cmake_minimum_required(VERSION 3.24)
+cmake_minimum_required(VERSION 3.24)
+cmake_minimum_required(VERSION 3.24...3.26)
